backend/django_app/requirements.txt
# Django Core
Django>=4.2,<5.0
django-environ>=0.10.0

# Database
psycopg2-binary>=2.9.9

# REST Framework
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
drf-spectacular>=0.27.0

# CORS
django-cors-headers>=4.3.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
redis>=5.0.0
django-redis>=5.4.0

# Email Service
resend>=0.7.0

# AI/LLM Services (Optional)
openai>=1.0.0
anthropic>=0.7.0

# Payment Processing
stripe>=7.0.0

# MFA/TOTP
pyotp>=2.9.0

# Scheduled Jobs (grace period, auto-downgrade, renewals)
django-apscheduler>=0.7.0

# Development (optional)
django-extensions>=3.2.3
django-filter>=23.3

# Monitoring
prometheus-client>=0.19.0
gunicorn>=21.2.0

# Security & Hashing
argon2-cffi>=23.1.0
PyJWT>=2.8.0
cryptography>=41.0.0

# Testing
pytest>=7.4.0
pytest-django>=4.7.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
factory-boy>=3.3.0

